Fix homography matrix precision

+++ HomographyLayer.cpp
@@ -24,12 +24,12 @@ void HomographyLayer::initialize(const std::vector<cv::KeyPoint> &points, const 
 
         // Find homography, identity if fails
         if (last_frame_matched.empty()) {
-            homography = cv::Mat::eye(3, 3, CV_32F);
+            homography = cv::Mat::eye(3, 3, CV_64F);
         } else {
             homography = cv::findHomography(last_frame_matched, new_frame_matched, cv::RANSAC, 0.1);
         }
         if (homography.empty()) {
-            homography = cv::Mat::eye(3, 3, CV_32F);
+            homography = cv::Mat::eye(3, 3, CV_64F);
         }
     }
